Optimize sir_list_del_item
We previously did a best effort attempt to see if the sir item exists
in any of the global lists, and if not, it must already be in the list
we wanted to add it to. This sucked hard...

This is what perf report said:
   Overhead  Command  Shared Object     Symbol
   ........  .......  ................  ..........................
   75.84%    main     main              [.] sir_list_del_item

Ouch, close to 76% runtime spent in searching for an item that is most
likely not in there. Let's use state variable to track the current
list and guard the extraction-insertion with a conditional for node's
state.